+++ Cargo.toml
@@ -5,22 +5,21 @@ members = [
     "adapters/avail",
     "adapters/risc0",
     "adapters/celestia",
+    "adapters/mock-da",
+    "adapters/mock-zkvm",
     "examples/const-rollup-config",
     "examples/demo-simple-stf",
     "examples/demo-rollup",
     "examples/simple-nft-module",
-
     "full-node/db/sov-db",
     "full-node/sov-sequencer",
     "full-node/sov-ethereum",
     "full-node/sov-ledger-rpc",
     "full-node/sov-stf-runner",
-
     "utils/zk-cycle-macros",
     "utils/zk-cycle-utils",
     "utils/bashtestmd",
     "utils/rng-da-service",
-
     "module-system/sov-cli",
     "module-system/sov-modules-stf-template",
     "module-system/sov-modules-rollup-template",
