+++ adapters/mock-da/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "sov-mock-da"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+publish = false
+
+[dependencies]
+anyhow = { workspace = true }
+async-trait = { workspace = true }
+borsh = { workspace = true }
+bytes = { workspace = true }
+serde = { workspace = true }
+hex = { workspace = true }
+sha2 = { workspace = true }
+tokio = { workspace = true, optional = true }
+
+sov-rollup-interface = { path = "../../rollup-interface", version = "0.3" }
+
+
+[features]
+default = []
+native = [
+    "dep:tokio",
+    "sov-rollup-interface/native"
+]
