+++ adapters/mock-da/src/lib.rs
@@ -1,0 +1,9 @@
+#[cfg(feature = "native")]
+mod service;
+mod types;
+mod validity_condition;
+
+#[cfg(feature = "native")]
+pub use service::*;
+pub use types::*;
+pub use validity_condition::*;
