+++ adapters/mock-da/src/types/mod.rs
@@ -1,11 +1,107 @@
+mod address;
+
+pub use address::{MockAddress, MOCK_SEQUENCER_DA_ADDRESS};
+use borsh::{BorshDeserialize, BorshSerialize};
 use bytes::Bytes;
 use serde::{Deserialize, Serialize};
-
-use crate::da::{BlobReaderTrait, CountedBufReader, DaSpec, DaVerifier};
-use crate::mocks::{
-    MockAddress, MockBlockHeader, MockDaSpec, MockDaVerifier, MockHash, MockValidityCond,
+use sov_rollup_interface::da::{
+    BlobReaderTrait, BlockHashTrait, BlockHeaderTrait, CountedBufReader, DaSpec, DaVerifier, Time,
 };
-use crate::services::da::SlotData;
+use sov_rollup_interface::services::da::SlotData;
+
+use crate::validity_condition::MockValidityCond;
+
+const JAN_1_2023: i64 = 1672531200;
+
+/// A mock hash digest.
+#[derive(
+    Clone,
+    Copy,
+    Debug,
+    PartialEq,
+    Eq,
+    serde::Serialize,
+    serde::Deserialize,
+    BorshDeserialize,
+    BorshSerialize,
+)]
+pub struct MockHash(pub [u8; 32]);
+
+impl AsRef<[u8]> for MockHash {
+    fn as_ref(&self) -> &[u8] {
+        &self.0
+    }
+}
+
+impl From<[u8; 32]> for MockHash {
+    fn from(value: [u8; 32]) -> Self {
+        Self(value)
+    }
+}
+
+impl From<MockHash> for [u8; 32] {
+    fn from(value: MockHash) -> Self {
+        value.0
+    }
+}
+
+impl BlockHashTrait for MockHash {}
+
+/// A mock block header used for testing.
+#[derive(Serialize, Deserialize, PartialEq, core::fmt::Debug, Clone, Copy)]
+pub struct MockBlockHeader {
+    /// The hash of the previous block.
+    pub prev_hash: MockHash,
+    /// The hash of this block.
+    pub hash: MockHash,
+    /// The height of this block
+    pub height: u64,
+}
+
+impl Default for MockBlockHeader {
+    fn default() -> Self {
+        Self {
+            prev_hash: MockHash([0u8; 32]),
+            hash: MockHash([1u8; 32]),
+            height: 0,
+        }
+    }
+}
+
+impl BlockHeaderTrait for MockBlockHeader {
+    type Hash = MockHash;
+
+    fn prev_hash(&self) -> Self::Hash {
+        self.prev_hash
+    }
+
+    fn hash(&self) -> Self::Hash {
+        self.hash
+    }
+
+    fn height(&self) -> u64 {
+        self.height
+    }
+
+    fn time(&self) -> Time {
+        Time::from_secs(JAN_1_2023 + (self.height as i64) * 15)
+    }
+}
+
+/// A [`crate::da::DaSpec`] suitable for testing.
+#[derive(serde::Serialize, serde::Deserialize, Debug, PartialEq, Eq)]
+pub struct MockDaSpec;
+
+/// The configuration for mock da
+#[derive(Debug, Clone, PartialEq, serde::Deserialize, serde::Serialize)]
+pub struct MockDaConfig {
+    /// The address to use to "submit" blobs on the mock da layer
+    pub sender_address: MockAddress,
+}
+
+#[derive(Clone, Default)]
+/// DaVerifier used in tests.
+pub struct MockDaVerifier {}
 
 #[derive(
     Debug,
@@ -28,7 +124,7 @@ impl MockBlob {
     pub fn new(data: Vec<u8>, address: MockAddress, hash: [u8; 32]) -> Self {
         Self {
             address,
-            data: CountedBufReader::new(bytes::Bytes::from(data)),
+            data: CountedBufReader::new(Bytes::from(data)),
             hash,
         }
     }
