+++ adapters/mock-zkvm/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "sov-mock-zkvm"
+authors = { workspace = true }
+description = "Mock zkVM implementation"
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+version = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = { workspace = true }
+borsh = { workspace = true }
+bincode = { workspace = true }
+serde = { workspace = true }
+sov-rollup-interface = { path = "../../rollup-interface", version = "0.3" }
+
+[features]
+default = []
+native = []
