+++ adapters/mock-zkvm/src/lib.rs
@@ -1,10 +1,9 @@
+use std::io::Write;
+
 use anyhow::ensure;
 use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
-
-use crate::maybestd::io;
-use crate::maybestd::vec::Vec;
-use crate::zk::Matches;
+use sov_rollup_interface::zk::Matches;
 
 /// A mock commitment to a particular zkVM program.
 #[derive(Debug, Clone, PartialEq, Eq, BorshDeserialize, BorshSerialize, Serialize, Deserialize)]
@@ -29,7 +28,7 @@ pub struct MockProof<'a> {
 
 impl<'a> MockProof<'a> {
     /// Serializes a proof into a writer.
-    pub fn encode(&self, mut writer: impl io::Write) {
+    pub fn encode(&self, mut writer: impl Write) {
         writer.write_all(&self.program_id.0).unwrap();
         let is_valid_byte = if self.is_valid { 1 } else { 0 };
         writer.write_all(&[is_valid_byte]).unwrap();
@@ -60,7 +59,7 @@ impl<'a> MockProof<'a> {
 /// A mock implementing the zkVM trait.
 pub struct MockZkvm;
 
-impl crate::zk::Zkvm for MockZkvm {
+impl sov_rollup_interface::zk::Zkvm for MockZkvm {
     type CodeCommitment = MockCodeCommitment;
 
     type Error = anyhow::Error;
@@ -78,29 +77,16 @@ impl crate::zk::Zkvm for MockZkvm {
         Ok(proof.log)
     }
 
-    #[cfg(feature = "std")]
     fn verify_and_extract_output<
-        Add: crate::RollupAddress,
-        Da: crate::da::DaSpec,
-        Root: Serialize + serde::de::DeserializeOwned,
+        Add: sov_rollup_interface::RollupAddress,
+        Da: sov_rollup_interface::da::DaSpec,
+        Root: serde::Serialize + serde::de::DeserializeOwned,
     >(
         serialized_proof: &[u8],
         code_commitment: &Self::CodeCommitment,
-    ) -> Result<crate::zk::StateTransition<Da, Add, Root>, Self::Error> {
+    ) -> Result<sov_rollup_interface::zk::StateTransition<Da, Add, Root>, Self::Error> {
         let output = Self::verify(serialized_proof, code_commitment)?;
         Ok(bincode::deserialize(output)?)
-    }
-
-    #[cfg(not(feature = "std"))]
-    fn verify_and_extract_output<
-        Add: crate::RollupAddress,
-        Da: crate::da::DaSpec,
-        Root: Serialize + serde::de::DeserializeOwned,
-    >(
-        _serialized_proof: &[u8],
-        _code_commitment: &Self::CodeCommitment,
-    ) -> Result<crate::zk::StateTransition<Da, Add, Root>, Self::Error> {
-        todo!("the current version of bincode doesn't support no-std; however, the next version is scheduled to")
     }
 }
