+++ examples/demo-rollup/Cargo.toml
@@ -13,9 +13,8 @@ default-run = "sov-demo-rollup"
 
 [dependencies]
 # non-optional dependencies
-sov-celestia-adapter = { path = "../../adapters/celestia", features = [
-	"native",
-] }
+sov-celestia-adapter = { path = "../../adapters/celestia", features = ["native"] }
+sov-mock-da = { path = "../../adapters/mock-da", features = ["native"] }
 const-rollup-config = { path = "../const-rollup-config" }
 sov-stf-runner = { path = "../../full-node/sov-stf-runner", features = ["native"] }
 sov-rollup-interface = { path = "../../rollup-interface", features = ["native"] }
@@ -25,7 +24,7 @@ sov-modules-stf-template = { path = "../../module-system/sov-modules-stf-templat
 sov-modules-api = { path = "../../module-system/sov-modules-api", features = ["native"] }
 sov-nft-module = { path = "../../module-system/module-implementations/sov-nft-module" }
 demo-stf = { path = "./stf", features = ["native"] }
-sov-ledger-rpc = { path = "../../full-node/sov-ledger-rpc", features = ["server",] }
+sov-ledger-rpc = { path = "../../full-node/sov-ledger-rpc", features = ["server", ] }
 risc0 = { path = "./provers/risc0" }
 borsh = { workspace = true, features = ["bytes"] }
 async-trait = { workspace = true }
