+++ examples/demo-rollup/benches/node/rollup_coarse_measure.rs
@@ -12,12 +12,12 @@ use demo_stf::runtime::Runtime;
 use prettytable::Table;
 use prometheus::{Histogram, HistogramOpts, Registry};
 use sov_db::ledger_db::{LedgerDB, SlotCommit};
+use sov_mock_da::{MockBlock, MockBlockHeader};
 use sov_modules_api::default_context::DefaultContext;
 use sov_modules_stf_template::kernels::basic::BasicKernel;
 use sov_modules_stf_template::{AppTemplate, TxEffect};
 use sov_risc0_adapter::host::Risc0Verifier;
 use sov_rng_da_service::{RngDaService, RngDaSpec};
-use sov_rollup_interface::mocks::{MockBlock, MockBlockHeader};
 use sov_rollup_interface::services::da::DaService;
 use sov_rollup_interface::stf::StateTransitionFunction;
 use sov_rollup_interface::storage::StorageManager;
