+++ examples/demo-rollup/provers/risc0/guest-mock/src/bin/mock_da.rs
@@ -1,10 +1,11 @@
 #![no_main]
 use demo_stf::runtime::Runtime;
 use demo_stf::AppVerifier;
+use sov_mock_da::MockDaVerifier;
 use sov_modules_api::default_context::ZkDefaultContext;
-use sov_modules_stf_template::{kernels::basic::BasicKernel, AppTemplate};
+use sov_modules_stf_template::kernels::basic::BasicKernel;
+use sov_modules_stf_template::AppTemplate;
 use sov_risc0_adapter::guest::Risc0Guest;
-use sov_rollup_interface::mocks::MockDaVerifier;
 use sov_state::ZkStorage;
 
 risc0_zkvm::guest::entry!(main);
