+++ examples/demo-rollup/src/main.rs
@@ -4,8 +4,8 @@ use anyhow::Context as _;
 use clap::Parser;
 use demo_stf::genesis_config::GenesisPaths;
 use sov_demo_rollup::{CelestiaDemoRollup, MockDemoRollup};
+use sov_mock_da::MockDaConfig;
 use sov_modules_rollup_template::{Rollup, RollupProverConfig, RollupTemplate};
-use sov_rollup_interface::mocks::MockDaConfig;
 use sov_stf_runner::{from_toml_path, RollupConfig};
 use tracing::log::debug;
 use tracing_subscriber::prelude::*;
