+++ full-node/sov-ethereum/README.md
@@ -1,0 +1,3 @@
+# sov-ethereum
+
+Crate that provides a thin wrapper for Sovereign SDK Rollup to have same interface as Ethereum RPC.
