+++ module-system/module-implementations/sov-attester-incentives/src/tests/helpers.rs
@@ -1,13 +1,13 @@
 use jmt::proof::SparseMerkleProof;
 use sov_bank::{BankConfig, TokenConfig};
+use sov_mock_da::{
+    MockBlock, MockBlockHeader, MockDaSpec, MockValidityCond, MockValidityCondChecker,
+};
+use sov_mock_zkvm::{MockCodeCommitment, MockZkvm};
 use sov_modules_api::default_context::DefaultContext;
 use sov_modules_api::hooks::SlotHooks;
 use sov_modules_api::utils::generate_address;
 use sov_modules_api::{Address, Genesis, Spec, ValidityConditionChecker, WorkingSet};
-use sov_rollup_interface::mocks::{
-    MockBlock, MockBlockHeader, MockCodeCommitment, MockDaSpec, MockValidityCond,
-    MockValidityCondChecker, MockZkvm,
-};
 use sov_state::storage::{NativeStorage, Storage, StorageProof};
 use sov_state::{DefaultStorageSpec, ProverStorage};
