+++ module-system/module-implementations/sov-blob-storage/tests/blob_storage_tests.rs
@@ -1,8 +1,8 @@
 use sov_blob_storage::BlobStorage;
 use sov_chain_state::{ChainState, ChainStateConfig};
+use sov_mock_da::{MockAddress, MockBlob, MockDaSpec};
 use sov_modules_api::default_context::DefaultContext;
 use sov_modules_api::{Module, WorkingSet};
-use sov_rollup_interface::mocks::{MockAddress, MockBlob, MockDaSpec};
 use sov_state::ProverStorage;
 
 type C = DefaultContext;
