+++ module-system/module-implementations/sov-chain-state/tests/all_tests.rs
@@ -1,10 +1,10 @@
 use sov_chain_state::{ChainState, ChainStateConfig, StateTransitionId, TransitionInProgress};
+use sov_mock_da::{MockBlock, MockBlockHeader, MockDaSpec, MockValidityCond};
 use sov_modules_api::da::BlockHeaderTrait;
 use sov_modules_api::default_context::DefaultContext;
 use sov_modules_api::hooks::SlotHooks;
+use sov_modules_api::storage::StorageManager;
 use sov_modules_api::{Genesis, WorkingSet};
-use sov_rollup_interface::mocks::{MockBlock, MockBlockHeader, MockDaSpec, MockValidityCond};
-use sov_rollup_interface::storage::StorageManager;
 use sov_state::config::Config;
 use sov_state::storage_manager::ProverStorageManager;
 use sov_state::{DefaultStorageSpec, Storage};
