+++ module-system/module-implementations/sov-sequencer-registry/Cargo.toml
@@ -14,7 +14,7 @@ resolver = "2"
 
 [dev-dependencies]
 sov-sequencer-registry = { path = ".", features = ["native"] }
-sov-rollup-interface = { path = "../../../rollup-interface", version = "0.3", features = ["mocks"] }
+sov-mock-da = { path = "../../../adapters/mock-da" }
 tempfile = { workspace = true }
 
 [dependencies]
