+++ module-system/sov-cli/Cargo.toml
@@ -30,4 +30,4 @@ jsonrpsee = { workspace = true, features = ["client"] }
 [dev-dependencies]
 tempfile = { workspace = true }
 demo-stf = { path = "../../examples/demo-rollup/stf", features = ["native"] }
-sov-rollup-interface = { path = "../../rollup-interface", features = ["mocks"] }
+sov-mock-da = { path = "../../adapters/mock-da" }
