+++ module-system/sov-cli/tests/transactions.rs
@@ -3,9 +3,9 @@ use std::path::{Path, PathBuf};
 use demo_stf::runtime::{Runtime, RuntimeCall, RuntimeSubcommand};
 use sov_cli::wallet_state::WalletState;
 use sov_cli::workflows::transactions::{ImportTransaction, TransactionWorkflow};
+use sov_mock_da::MockDaSpec;
 use sov_modules_api::cli::{FileNameArg, JsonStringArg};
 use sov_modules_api::default_context::DefaultContext;
-use sov_rollup_interface::mocks::MockDaSpec;
 
 type Da = MockDaSpec;
