+++ module-system/utils/sov-data-generators/src/lib.rs
@@ -1,14 +1,11 @@
 use std::rc::Rc;
 
 use borsh::ser::BorshSerialize;
+use sov_mock_da::{MockAddress, MockBlob, MockDaSpec};
 use sov_modules_api::transaction::Transaction;
 pub use sov_modules_api::EncodeCall;
-use sov_modules_api::{Context, Module, Spec};
+use sov_modules_api::{Context, DaSpec, Module, RollupAddress, Spec};
 use sov_modules_stf_template::{Batch, RawTx, SequencerOutcome, TxEffect};
-use sov_rollup_interface::da::DaSpec;
-use sov_rollup_interface::mocks::{MockAddress, MockBlob, MockDaSpec};
-use sov_rollup_interface::stf::BatchReceipt;
-use sov_rollup_interface::RollupAddress;
 
 pub mod bank_data;
 pub mod value_setter_data;
@@ -23,16 +20,17 @@ pub fn new_test_blob_from_batch(
     MockBlob::new(data, address, hash)
 }
 
-pub fn has_tx_events<A: RollupAddress>(
-    apply_blob_outcome: &BatchReceipt<SequencerOutcome<A>, TxEffect>,
-) -> bool {
-    let events = apply_blob_outcome
-        .tx_receipts
-        .iter()
-        .flat_map(|receipts| receipts.events.iter());
-
-    events.peekable().peek().is_some()
-}
+// pub fn has_tx_events<A: RollupAddress>(
+//     apply_blob_outcome: &BatchReceipt<SequencerOutcome<A>, TxEffect>,
+//     a: &[TransactionReceipt<A>],
+// ) -> bool {
+//     let events = apply_blob_outcome
+//         .tx_receipts
+//         .iter()
+//         .flat_map(|receipts| receipts.events.iter());
+//
+//     events.peekable().peek().is_some()
+// }
 
 /// A generic message object used to create transactions.
 pub struct Message<C: Context, Mod: Module> {
