+++ rollup-interface/Cargo.toml
@@ -45,7 +45,6 @@ proptest-derive = { workspace = true }
 default = ["std"]
 native = ["std", "tokio"]
 fuzzing = ["proptest", "proptest-derive", "sha2", "std"]
-mocks = ["sha2", "bytes/serde", "std"]
 std = [
     "anyhow/default",
     "bincode",
