+++ utils/rng-da-service/Cargo.toml
@@ -20,3 +20,4 @@ demo-stf = { path = "../../examples/demo-rollup/stf" }
 sov-bank = { path = "../../module-system/module-implementations/sov-bank" }
 sov-modules-api = { path = "../../module-system/sov-modules-api", features = ["native"] }
 sov-rollup-interface = { path = "../../rollup-interface", features = ["native"] }
+sov-mock-da = { path = "../../adapters/mock-da" }
